requirements.txt
# Core dependencies
telethon==1.40.0
fastapi==0.115.14
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic==2.9.2
pydantic-settings==2.10.1

# Database dependencies
redis==6.2.0
neo4j==5.28.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.41

# ML/NLP dependencies
spacy==3.8.7
numpy==1.26.4
pandas==2.3.0
transformers==4.45.2
sentence-transformers==3.3.1
torch>=2.6.0
scipy==1.14.1
regex>=2024.0.0
tokenizers>=0.20.0,<0.21.0
huggingface-hub>=0.16.0
safetensors>=0.4.0
numpy>=1.24.0,<2.0.0

# Task queue
celery==5.5.3
flower==2.0.1

# Utilities
pyyaml==6.0.2
python-json-logger==3.2.1
httpx==0.28.1
aiofiles==24.1.0

# Development dependencies
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
black==24.10.0
flake8==7.3.0
mypy==1.16.1
ipython==8.31.0

# Code quality and CI/CD tools
pre-commit==4.2.0
isort==6.0.1
bandit==1.8.5
safety==3.5.2
ruff==0.12.1

# Additional flake8 plugins
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-docstrings==1.7.0
flake8-import-order==0.19.2

# Type stubs
types-redis==4.6.0.20241004
types-pyyaml==6.0.12.20241230